wip: separate files for states and commands
