+++ logic/handlers/clear.py
@@ -1,0 +1,22 @@
+import logging
+
+from aiogram import F
+from aiogram.filters import Command
+from aiogram.fsm.context import FSMContext
+from aiogram.types import Message, ReplyKeyboardRemove
+
+from logic.handlers import router
+
+
+@router.message(Command("cancel"))
+@router.message(F.text.casefold() == "cancel")
+async def cancel_handler(message: Message, state: FSMContext) -> None:
+    current_state = await state.get_state()
+    if current_state is None:
+        return
+    logging.info("Cancelling state %r", current_state)
+    await state.clear()
+    await message.answer(
+        "Вы сбросили сохраненное состояние.",
+        reply_markup=ReplyKeyboardRemove(),
+    )
