+++ logic/states.py
@@ -1,0 +1,7 @@
+from aiogram.fsm.state import StatesGroup, State
+
+
+class PossibleStates(StatesGroup):
+    start = State()
+    tranlog = State()
+    open = State()
